Fix address_set and port_group usage in create_routed_lport scenario
When an address set is used in the acl match it shoule be used as
"ip4.src == $as", where as it's used as "ip4.src == as".

When a port group is used in the acl match it should be used as
"inport == @pg", where as it's used as "inport == "@pg"".

This patch fixes these.

Fixes: 8594ad8f9a5b("add port-group support to openshift workload")

Signed-off-by: Numan Siddique <nusiddiq@redhat.com>